refactor: Post Detail use iglistkit for post
