+++ spotcheck-ios/Controllers/Posts/FeedViewController.swift
@@ -280,13 +280,13 @@ extension FeedViewController: UICollectionViewDataSource, UICollectionViewDelega
     // MARK: - When Cell selected or clicked
     func collectionView(_ collectionView: UICollectionView, didSelectItemAt indexPath: IndexPath) {
         let post = posts[indexPath.row]
-        
+
         if indexPath.section == 1 {
             return
         }
-        
+
         let cell = collectionView.cellForItem(at: indexPath) as! FeedCell
-        
+
         if let player = cell.avPlayerViewController.player {
             player.pause()
             //player.play() TODO: figure how to programatticaly do full screen mode
@@ -294,7 +294,7 @@ extension FeedViewController: UICollectionViewDataSource, UICollectionViewDelega
             //window.rootViewController?.present(cell.avPlayerViewController, animated: true)
             //window.makeKeyAndVisible()
         }
-        
+
         viewPostHandler(exercisePost: post)
     }
 
@@ -372,13 +372,13 @@ private extension FeedViewController {
     //Renders the changes between self's posts[] and the arg's posts[]
     func diffedTableViewRenderer(argPosts: [ExercisePost]) {
         //new data comes in `argPosts`
-        let results = ListDiffPaths(fromSection: 0, toSection: 0, oldArray: self.posts, newArray: argPosts, option: .equality)
+//        let results = ListDiffPaths(fromSection: 0, toSection: 0, oldArray: self.posts, newArray: argPosts, option: .equality)
 
-        self.posts = argPosts // set arg data into exiting array before updating tableview
-        self.feedView.performBatchUpdates({
-            self.feedView.deleteItems(at: results.deletes)
-            self.feedView.insertItems(at: results.inserts)
-        })
+//        self.posts = argPosts // set arg data into exiting array before updating tableview
+//        self.feedView.performBatchUpdates({
+//            self.feedView.deleteItems(at: results.deletes)
+//            self.feedView.insertItems(at: results.inserts)
+//        })
       //TODO: Do fade animation (?) for deletes and automatic for insert
     }
 
@@ -421,7 +421,7 @@ private extension FeedViewController {
     }
 
     func viewPostHandler(exercisePost: ExercisePost) {
-        let postDetailViewController = PostDetailViewController.create(postId: exercisePost.id, diffedPostsDataClosure: self.diffedPostsHandler)
+        let postDetailViewController = PostDetailViewController.create(postId: exercisePost.id)
         self.navigationController?.pushViewController(postDetailViewController, animated: true)
     }
 }
